+++ src/bittytax/conv/parsers/exodus.py
@@ -58,9 +58,6 @@ def parse_exodus_v2(data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[P
             wallet=WALLET,
             note=row_dict["PERSONALNOTE"],
         )
-    elif row_dict["TYPE"] == "deposit (failed)":
-        # Skip failures
-        return
     elif row_dict["TYPE"] == "withdrawal":
         sell_quantity, sell_asset = _split_asset(row_dict["COINAMOUNT"])
 
@@ -74,6 +71,24 @@ def parse_exodus_v2(data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[P
             wallet=WALLET,
             note=row_dict["PERSONALNOTE"],
         )
+    elif row_dict["TYPE"] in ("deposit (failed)", "withdrawal (failed)"):
+        # Just spend the fee for failures
+        if fee_quantity:
+            if row_dict["PERSONALNOTE"]:
+                note = f'Failure ({row_dict["PERSONALNOTE"]})'
+            else:
+                note = "Failure"
+
+            data_row.t_record = TransactionOutRecord(
+                TrType.SPEND,
+                data_row.timestamp,
+                sell_quantity=Decimal(0),
+                sell_asset=fee_asset,
+                fee_quantity=fee_quantity,
+                fee_asset=fee_asset,
+                wallet=WALLET,
+                note=note,
+            )
     else:
         raise UnexpectedTypeError(parser.in_header.index("TYPE"), "TYPE", row_dict["TYPE"])
 
@@ -113,9 +128,6 @@ def parse_exodus_v1(data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[P
             wallet=WALLET,
             note=row_dict["PERSONALNOTE"],
         )
-    elif row_dict["TYPE"] == "deposit (failed)":
-        # Skip failures
-        return
     elif row_dict["TYPE"] == "withdrawal":
         data_row.tx_raw = TxRawPos(
             parser.in_header.index("OUTTXID"), tx_dest_pos=parser.in_header.index("TOADDRESS")
@@ -130,6 +142,25 @@ def parse_exodus_v1(data_row: "DataRow", parser: DataParser, **_kwargs: Unpack[P
             wallet=WALLET,
             note=row_dict["PERSONALNOTE"],
         )
+    elif row_dict["TYPE"] in ("deposit (failed)", "withdrawal (failed)"):
+        # Just spend the fee for failures
+        if fee_quantity:
+            if row_dict["PERSONALNOTE"]:
+                note = f'Failure ({row_dict["PERSONALNOTE"]})'
+            else:
+                note = "Failure"
+
+        if fee_quantity:
+            data_row.t_record = TransactionOutRecord(
+                TrType.SPEND,
+                data_row.timestamp,
+                sell_quantity=Decimal(0),
+                sell_asset=fee_asset,
+                fee_quantity=fee_quantity,
+                fee_asset=fee_asset,
+                wallet=WALLET,
+                note=note,
+            )
     elif row_dict["TYPE"] == "exchange":
         data_row.t_record = TransactionOutRecord(
             TrType.TRADE,